+++ com.onesignal.unity.core/Runtime/User/Models/IUserState.cs
@@ -32,14 +32,14 @@ namespace OneSignalSDK.User {
         /// <summary>
         /// The unique identifier for your OneSignal account. This will be an empty string until the
         /// user has been successfully logged in on the backend and assigned an ID.
-        /// Use [addObserver] to be notified when the [OnesignalId] has been successfully assigned.
+        /// Use [Changed] to be notified when the [OneSignalId] has been successfully assigned.
         /// </summary>
         string OneSignalId { get; }
         
         /// <summary>
-        /// The external identifier that you use to identify users. Use [addObserver] to be notified
-        /// when the [externalId] has been successfully assigned. This will be an empty string if no
-        /// external identifier has been assigned to the associated [OnesignalId].
+        /// The external identifier that you use to identify users. Use [Changed] to be notified
+        /// when the [ExternalId] has been successfully assigned. This will be an empty string if no
+        /// external identifier has been assigned to the associated [OneSignalId].
         /// </summary>
         string ExternalId { get; }
     }