+++ OneSignalExample/Assets/OneSignal/CHANGELOG.md
@@ -8,8 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - iOS crash when calling OneSignal.User.PushSubscription.Id and OneSignal.User.PushSubscription.Token when they are null.
 ### Changed
-- Add public getters for OnesignalId and ExternalId in the User namespace
-- Add public event handler OneSignal.User.Changed that fires when the OnesignalId or ExternalId changes
+- Add public getters for OneSignalId and ExternalId in the User namespace
+- Add public event handler OneSignal.User.Changed that fires when the OneSignalId or ExternalId changes
 - Updated included Android SDK to [5.1.4](https://github.com/OneSignal/OneSignal-Android-SDK/releases/tag/5.1.4)
 
 ## [5.0.6]